Encapsulate high-level behaviour at class level
